Use HGNC BioMart to recover entries without HGNC IDs
Take the NCBI/Entrez Gene IDs without HGNC IDs to HGNC BioMart and
map them to HGNC IDs. With this information, some Ensembl gene IDs are
also recovered.

Note that none of the HGNC symbols are updated from this recovery to
maintain the consistency of the gene symbols. So the symbols always
match the RefSeq database and related annotations of the peptide
database downloaded from DCC.

HGNC regularly updates the approved symbol names, one can always obtain
the latest gene symbols using the HGNC IDs, which are more stable.